// front-end/src/types/projetos.d.ts
export type IProjetos = {
    id: string,
    nome: string,
    impacto: string,
    objetivo: string,
    atividades: string,
    responsavel: string,
    valor: string,
    patrocinadores?: string,
    osc?: any,
    beneficiarios?: any,
    oSCId: string,
    enderecos: string,
    enderecosId: string
}